// src/types.ts
export type AtLeast<T, K extends keyof T> = Partial<T> & Pick<T, K>;
// ...
export enum EVariableType {
  Level = 'LEVEL',
  List = 'LIST',
  Text = 'TEXT',
}

export interface Variable {
  name?: string;
  type: EVariableType;
  supported_feature?: number;
}

export interface ListVariableOption {
  value: string;
  icon?: string;
  name?: string;
}

export interface ListVariable extends Variable {
  options: ListVariableOption[];
}